+++ CHANGELOG.md
@@ -9,11 +9,11 @@ https://github.com/coleifer/peewee/releases
 
 * Add `MySQLJSONField` (`playhouse.mysql_ext`) with `contains_any()` for the
   `JSON_OVERLAPS`/"match any" counterpart to `contains` for JSON arrays.
-* New declarative API for pre-fetching related instance (`Load()`). See
+* New declarative API for pre-fetching related instances (`Load()`). See
   [documentation](https://docs.peewee-orm.com/en/latest/peewee/relationships.html#avoiding-n-1-with-eager-loading).
   This replaces `prefetch()`, is more flexible and also supports options for
-  applying a row limit to sub-results, and an option to materialize the ID
-  list.
+  applying a row limit to sub-results, and a strategy that materializes the ID
+  list (in addition to SELECT IN and JOIN strategies).
 
 [View commits](https://github.com/coleifer/peewee/compare/4.1.0...master)